Apps migration

+++ pkg/controllers/managementuserlegacy/helm/controller.go
@@ -269,29 +269,31 @@ func (l *Lifecycle) Remove(obj *v3.App) (runtime.Object, error) {
 			return obj, err
 		}
 	}
-	tempDirs, err := createTempDir(obj)
-	if err != nil {
-		return obj, err
-	}
-	defer os.RemoveAll(tempDirs.FullPath)
-	tokenName, err := l.writeKubeConfig(obj, tempDirs.KubeConfigFull, true)
-	if err != nil {
-		return obj, err
-	}
-	defer func() {
-		if err := l.systemTokens.DeleteToken(tokenName); err != nil {
-			logrus.Errorf("cleanup for helm token [%s] failed, will not retry: %v", tokenName, err)
+	if !(obj.Annotations["cattle.io/skipUninstall"] == "true") {
+		tempDirs, err := createTempDir(obj)
+		if err != nil {
+			return obj, err
 		}
-	}()
-	// try three times and succeed
-	start := time.Second * 1
-	for i := 0; i < 3; i++ {
-		if err = helmDelete(tempDirs, obj); err == nil {
-			break
+		defer os.RemoveAll(tempDirs.FullPath)
+		tokenName, err := l.writeKubeConfig(obj, tempDirs.KubeConfigFull, true)
+		if err != nil {
+			return obj, err
 		}
-		logrus.Warn(err)
-		time.Sleep(start)
-		start *= 2
+		defer func() {
+			if err := l.systemTokens.DeleteToken(tokenName); err != nil {
+				logrus.Errorf("cleanup for helm token [%s] failed, will not retry: %v", tokenName, err)
+			}
+		}()
+		// try three times and succeed
+		start := time.Second * 1
+		for i := 0; i < 3; i++ {
+			if err = helmDelete(tempDirs, obj); err == nil {
+				break
+			}
+			logrus.Warn(err)
+			time.Sleep(start)
+			start *= 2
+		}
 	}
 	ns, err := l.NsClient.Get(obj.Spec.TargetNamespace, metav1.GetOptions{})
 	if err != nil && !errors.IsNotFound(err) {
